test: fix in browser
