+++ test/tcp-duplex-browser.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = cb => cb(new Error('Please skip this test in the browser!'))
