+++ test/util.js
@@ -5,8 +5,6 @@ const Connection = require('interface-connection').Connection
 const multiaddr = require('multiaddr')
 const pull = require('pull-stream')
 const assert = require('assert')
-const TCP = require('libp2p-tcp')
-const tcp = new TCP()
 
 const Duplex = () => {
   const d = _Duplex()
@@ -18,17 +16,7 @@ const Duplex = () => {
   })
 }
 
-const TCPDuplex = cb => {
-  const server = tcp.createListener()
-  server.listen(multiaddr('/ip4/127.0.0.1/tcp/0'))
-  server.once('listening', () => {
-    server.getAddrs((err, addr) => {
-      if (err) return cb(err)
-      const client = tcp.dial(addr[0])
-      server.on('connection', sclient => cb(null, client, sclient))
-    })
-  })
-}
+const TCPDuplex = require('./tcp-duplex')
 
 module.exports = {
   Duplex,
@@ -41,5 +29,6 @@ module.exports = {
     }
     assert.deepEqual(v, res)
     if (cb) cb()
-  })
+  }),
+  skipbrowser: it => process.toString() === '[object process]' ? it : it.skip
 }
