+++ test/zero.handshake.spec.js
@@ -2,8 +2,13 @@
 
 'use strict'
 
-const {/*Duplex,*/ TCPDuplex} = require('./util')
+const {Duplex, TCPDuplex, skipbrowser} = require('./util')
 const Protocol = require('../src').Zero
+
+const chai = require('chai')
+const dirtyChai = require('dirty-chai')
+const expect = chai.expect
+chai.use(dirtyChai)
 
 const protocol = new Protocol({}, {swarm: {}})
 
@@ -14,25 +19,26 @@ protocol.handle('ping', {}, {body: [
 const doServer = protocol.upgradeConn({isServer: true})
 const doClient = protocol.upgradeConn({isServer: false})
 
-/* it('can send pings', cb => { TODO: fix? sync stuff
-  const [client, server] = Duplex()
-  doServer(server, console.log)
-  doClient(client, (err, c) => {
-    if (err) return cb(err)
-    console.log(c)
-    c.cmd.ping({body: 'Pong!'}, cb)
-  })
-}) */
+describe('zero protocol', () => {
+  describe('handshake', () => {
+    it('can send pings', cb => {
+      const [client, server] = Duplex()
+      doServer(server, err => expect(err).to.not.exist())
+      doClient(client, (err, c) => {
+        if (err) return cb(err)
+        c.cmd.ping({body: 'Pong!'}, cb)
+      })
+    })
 
-
-it('can send pings via tcp', cb => {
-  TCPDuplex((err, client, server) => {
-    if (err) return cb(err)
-    doServer(server, console.log)
-    doClient(client, (err, c) => {
-      if (err) return cb(err)
-      console.log(c)
-      c.cmd.ping({body: 'Pong!'}, cb)
+    skipbrowser(it)('can send pings via tcp', cb => {
+      TCPDuplex((err, client, server) => {
+        if (err) return cb(err)
+        doServer(server, err => expect(err).to.not.exist())
+        doClient(client, (err, c) => {
+          if (err) return cb(err)
+          c.cmd.ping({body: 'Pong!'}, cb)
+        })
+      })
     })
   })
 })
